Move item disenchanting static req. checks to server start.
(based on commit e40f9ef)

+++ src/game/ItemHandler.cpp
@@ -409,11 +409,11 @@ void WorldSession::HandleItemQuerySingleOpcode( WorldPacket & recv_data )
             data << pProto->Socket[s].Color;
             data << pProto->Socket[s].Content;
         }
-        data << pProto->socketBonus;
-        data << pProto->GemProperties;
-        data << pProto->RequiredDisenchantSkill;
-        data << pProto->ArmorDamageModifier;
-        data << pProto->Duration;                           // added in 2.4.2.8209, duration (seconds)
+        data << uint32(pProto->socketBonus);
+        data << uint32(pProto->GemProperties);
+        data << int32(pProto->RequiredDisenchantSkill);
+        data << float(pProto->ArmorDamageModifier);
+        data << uint32(pProto->Duration);                   // added in 2.4.2.8209, duration (seconds)
         SendPacket( &data );
     }
     else
